refactor: split AI Builder chat panel
